chore(setup): rename dialog package
